feat: add tracer skeleton for create_wallet use case (#9)
This stubs out lightweight implementations of everything needed for this
use case that roughly follows the architecture in the design sketch.
This includes all wiring and unit tests to make sure future code has a
good launching point.

The `create_wallet` implementation is a no-op and returns an empty
response object. The implementation will follow after this.