feat: add openapi codegen, frontrunner-api (#24)
Add script + docs for script on how to generate openapi. Add
frontrunner-api as first case.

NOTE: there's a chance that not putting `BUILD` files in the generated
code is wrong. However, with them in place, they'll trigger all the
formatters and checkers, which just flat out fail. It's near impossible
to configure some of these to ignore correctly, so it's easier just to
ignore them at the pants level. Hopefully, this is the right call.