feat: include local order hashes, allow order hash correlation (#105)
This adds locally-computed order hashes so that devs can detect _which_
order in a transaction failed, if any.

Even if a transaction succeeds, some orders within may fail. When
retrieving the transaction, the "failed" orders are expressed as a hash.
The hash can only be retrieved (for failed transactions) by local
computation. It's non-trivial to get these, so it's easies to return
them alongside order creation. Then, on `get_transaction`, you can see
which order(s) failed by seeing which order's hashes are in the
`get_transaction` hashes.